pyproject.toml
[tool.poetry]
name = "WszebiPatientPortal"
version = "0.1.0"
description = "App for booking appointments for doctor"
authors = ["wszeborj <wszeborj@o2.pl>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.8"
django = "^5.1.3"
django-environ = "^0.11.2"
django-bootstrap-datepicker-plus = "^5.0.5"
pillow = "^11.0.0"
django-bootstrap4 = "^24.4"
django-extensions = "^3.2.3"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.10"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.1"
flower = "^2.0.1"
pytest = "^8.3.5"
django-filter = "^25.1"
django-silk = "^5.4.0"
django-debug-toolbar = "^5.2.0"
safety = "^3.6.0"
coverage = "^7.9.2"
factory-boy = "^3.3.3"



[tool.poetry.group.dev.dependencies]
poetry = "^1.8.4"
pre-commit = "^4.0.1"
sentry-sdk = "^2.19.0"
bandit = "^1.8.0"
ruff = "^0.8.1"
icecream = "^2.1.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
